# contracts/openapi/horizon-api.yaml
openapi: 3.1.0
info:
  title: Horizon Public API
  version: 1.0.0
  description: |
    Публичный API платформы Horizon — обнаружение зарождающихся научно-технологических
    трендов (слабых сигналов) в открытых источниках.

    Контракт **first-class**: спецификация является источником истины, реализация проверяется
    контрактными тестами (`spring-boot-starter-web` + `swagger-request-validator`, Schemathesis для Python).

    Все ошибки — RFC 9457 `application/problem+json` (ADR-0013).
    Входа нет: организация одна, и API открыт так же, как интерфейс. Служебные вызовы между
    сервисами защищены общим секретом (`X-Internal-Token`) и в этот контракт не входят.
  contact:
    name: Horizon Team
    url: https://horizon.dev
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Локальная среда (gateway)
  - url: https://horizon.example.bank
    description: Промышленная среда
...
paths:
...
  /api/v1/reports/{reportId}:
    get:
      tags: [Reports]
      summary: Полный отчёт по трендам
      operationId: getReport
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Отчёт
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrendReport' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
  /api/v1/reports/{reportId}/delta:
    get:
      tags: [Reports]
      summary: Что изменилось с предыдущей версии отчёта
      description: >-
        Состав ТОП, размеченный на появившиеся, выбывшие и оставшиеся темы, с изменением ранга и
        балла у оставшихся. Отдельный ресурс, а не поле отчёта: дельта нужна при повторном чтении,
        и вкладывать её значило бы поднимать вторую версию из хранилища на каждый запрос отчёта.


        Темы сопоставляются по `trendKey`, а не по заголовку: заголовок — это подпись кластера, и
        она законно меняется при том же предмете, поэтому сравнение по нему выдало бы улучшение
        названия за событие в реальности.
      operationId: getReportDelta
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Дельта либо причина, по которой её нет
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportDelta' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
  /api/v1/reports/{reportId}/explain:
    get:
      tags: [Reports]
      summary: Почему темы нет в отчёте
      description: >-
        Прослеживает указанные термины через конвейер анализа и сообщает, на какой стадии каждый из
        них был отброшен и по какой причине. Отвечает на вопрос «я знаю, что эта технология
        существует, — почему её нет в отчёте?», который иначе остаётся без ответа и разрушает
        доверие к ранжированию.


        Ответ строится повторным прогоном анализа на том же срезе корпуса — это минуты, поэтому
        запрос не ждёт движка: пока прогон идёт, ответ — 202 с `Retry-After`, и клиент повторяет тот
        же запрос, пока не получит 200. Одинаковые вопросы сводятся к одному прогону. Число терминов
        ограничено пятью.
      operationId: explainReportTerms
      parameters:
        - $ref: '#/components/parameters/ReportId'
        - name: term
          in: query
          required: true
          description: Термин; параметр можно повторить, но не более пяти раз.
          schema:
            type: array
            minItems: 1
            maxItems: 5
            items: { type: string }
      responses:
        '200':
          description: Что произошло с каждым термином
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TermExplanation' }
        '202':
          description: Прогон запущен или ещё идёт; повторите тот же запрос через `Retry-After` секунд
          headers:
            Retry-After:
              description: Через сколько секунд спросить снова
              schema: { type: integer, minimum: 1 }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TermExplanationPending' }
        '400': { $ref: '#/components/responses/ValidationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
        '502':
          description: Движок анализа не ответил на запрос трассировки
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/reports/{reportId}/trends/{trendKey}:
    get:
      tags: [Reports]
      summary: Один тренд с полной детализацией
      operationId: getReportTrend
      parameters:
        - $ref: '#/components/parameters/ReportId'
        - name: trendKey
          in: path
          required: true
          schema: { type: string, maxLength: 160 }
      responses:
        '200':
          description: Тренд
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RankedTrend' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
  /api/v1/reports/{reportId}/export:
    get:
      tags: [Reports]
      summary: Экспорт отчёта
      description: >-
        Три формата под три разных читателя: JSON для скрипта, CSV для Excel, Markdown для человека,
        который переносит отчёт в записку комитету. Последний нужен потому, что именно на этом
        переносе оговорки терялись: вручную переносят выводы, а не сноски.


        Во всех форматах отсутствуют личные пометки аналитика: файл передают дальше, а дословный
        комментарий, написанный для себя, в чужих руках — это не то, что человек соглашался
        опубликовать. Документ датируется анализом, а не моментом выгрузки, поэтому две выгрузки
        одного отчёта дают одинаковые байты.
      operationId: exportReport
      parameters:
        - $ref: '#/components/parameters/ReportId'
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [json, csv, markdown] }
      responses:
        '200':
          description: Файл экспорта
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrendReport' }
            text/markdown:
              # Записка: заголовок с версией и методологией, блок оговорок, затем темы по порядку
              # рейтинга — у каждой определение, проблема, польза, балл с уверенностью и ссылки на
              # источники. Оговорки стоят ДО тем намеренно: сноска в конце отбрасывается при
              # копировании первой половины документа, и до комитета доезжает только та оговорка,
              # которую нельзя не увидеть.
              schema: { type: string }
            text/csv:
              # Файл состоит из двух блоков: портрет направления парами «ключ;значение», пустая
              # строка, затем таблица трендов. Первая строка намеренно не является заголовком
              # таблицы: BOM выше уже объявляет этот формат тем, который открывают люди, машинам
              # служит JSON-экспорт, а оговорки (доля тем на тонкой доказательной базе, недоступные
              # источники) по BR-A17 не могут быть опущены — и под пятнадцатью строками они спрятаны.
              schema: { type: string }
        '404': { $ref: '#/components/responses/NotFoundProblem' }

  # ───────────────────────────── Feedback ─────────────────────────────
...
components:
  parameters:
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    ValidationProblem:
      description: Ошибка валидации
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFoundProblem:
      description: Не найдено
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, examples: ['https://horizon.dev/problems/validation-error'] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
        timestamp: { type: string, format: date-time }
        retryable: { type: boolean }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
              rejectedValue: {}
...
    RankedTrend:
      type: object
      required:
        [rank, trendKey, title, definition, motivation, assessment,
         lifecycleStage, firstMentionYear, totalDocuments, timeline, evidence]
      properties:
        rank: { type: integer, minimum: 1 }
        trendKey: { type: string, maxLength: 160 }
        title: { type: string, maxLength: 200 }
        definition: { type: string }
        motivation: { $ref: '#/components/schemas/Motivation' }
        caseExample: { $ref: '#/components/schemas/CaseExample' }
        assessment: { $ref: '#/components/schemas/EmergenceAssessment' }
        directionShare:
          type: number
          minimum: 0
          maximum: 1
          description: |
            Доля документов темы, отнесённых источниками к направлению отчёта.

            Величина, на которой держится отбор темы в отчёт. У настоящих тем направления она около
            двух третей; тема с седьмой частью попадает в отчёт потому, что для узкого направления
            планка опускается почти до нуля. Статистика их не разделяет — человек разделяет сразу,
            если величину показать. Отсутствует, когда отбор шёл не по предметным кодам.
        lifecycleStage: { $ref: '#/components/schemas/LifecycleStage' }
        firstMentionYear:
          type: integer
          minimum: 1900
          maximum: 2100
          description: >-
            Первый достоверный год темы по BRULE-2: не меньше двух документов из не меньше чем двух
            организаций. Это не год первого упоминания вообще — одиночные упоминания бывают раньше:
            по эталонному корпусу у 71 темы из 90, у 21 на два года и больше. Имя поля оставлено
            прежним намеренно: переименование сломало бы потребителей ради названия, а различие
            несут это описание и все подписи в интерфейсе.
        totalDocuments: { type: integer, minimum: 1 }
        burst:
          type: object
          properties:
            startPeriod: { type: [string, 'null'] }
            weight: { type: [number, 'null'] }
        timeline:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/TimelinePoint' }
        evidence:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Evidence' }
        lowCredibilityOnly:
          type: boolean
          default: false
          description: >-
            Все источники темы низкой доверенности. ТЗ: такие сведения «не должны быть единственным
            основанием для включения технологии в итоговую выдачу… либо сопровождаться отметкой о
            пониженной доверенности». Первую половину закрывает правило движка, отсекающее тему без
            научно-технической основы вовсе; вторую — эта отметка.
        localization: { $ref: '#/components/schemas/Localization' }
        explanation:
          type: array
          maxItems: 8
          description: >-
            Почему тема — слабый сигнал и почему у неё такая уверенность: строки «заголовок —
            пояснение», собранные движком из голосов экспертной стадии, вкладов в балл, расчёта
            уверенности и источников карточки. Пусто у отчётов, выпущенных до появления объяснения.
          items: { $ref: '#/components/schemas/ExplanationItem' }
        feedback: { $ref: '#/components/schemas/TrendFeedbackView' }
...
    Coverage:
      type: object
      required: [documentsAnalyzed, sourcesUsed, partial]
      properties:
        documentsAnalyzed: { type: integer }
        candidatesEvaluated: { type: integer }
        sourcesUsed: { type: array, items: { type: string } }
        unavailableSources: { type: array, items: { type: string } }
        partial: { type: boolean }
        # Оговорка того же рода, что partial, но о другом отказе. partial говорит «мы нашли
        # меньше, чем хотели»; это поле — «мы не поняли вопроса»: движок не соотнёс запрошенное
        # направление со словарём предметных кодов, которым размечен корпус, и выбор тем не
        # зависит от заданного вопроса. Отчёт при этом полон, посчитан и правдоподобен.
        # Необязательное: отчёты, выпущенные до появления поля, читаются как распознанные —
        # оговорка, срабатывающая не по делу, обесценивает настоящие.
        directionRecognized: { type: boolean, default: true }
        # Непусто только при directionRecognized=false. Без подсказки красная плашка «направление
        # не распознано» — тупик: разница между работающим продуктом и бесполезным здесь часто в
        # одном слове, которое аналитику неоткуда узнать.
        directionSuggestions:
          type: array
          maxItems: 3
          items: { type: string }
        # Сколько тем скрыто по пометке аналитика «это не технология». Отбор происходит до ТОП-N,
        # поэтому освободившееся место занял следующий кандидат — и без этого числа отчёт выглядел
        # бы так, будто скрытого не было вовсе. Молчаливый фильтр производит уверенность вместо
        # сведений: аналитик обязан видеть, что список неполон по его же решению.
        suppressedByAnalyst: { type: integer, minimum: 0, default: 0 }
        # Корпус обрезан пределом профиля до анализа: часть литературы направления не
        # рассматривалась вовсе. Отличается от partial («источник был недоступен») и от «тем найдено
        # меньше запрошенных» («методология столько и нашла»): три разные причины неполноты, и
        # читателю они говорят разное.
        corpusTruncated: { type: boolean, default: false }
        # Четвёртая оговорка того же рода, и ТЗ требует её прямо: «в веб-интерфейсе должны
        # отображаться … причины исключения зрелых технологий или нерелевантных кандидатов при
        # наличии такой информации», и отдельно — «должна быть продемонстрирована логика исключения
        # зрелых трендов, массово внедренных технологий, отраслевых стандартов, маркетингового хайпа
        # и информационного шума». Примеры идут вместе со счётчиком: число без имён проверить
        # нельзя — «отсеяно 2400 кандидатов» одинаково читается и там, где отсеяли мусор, и там,
        # где отсеяли тему, которую аналитик искал.
        exclusions:
          type: array
          items:
            type: object
            required: [code, reason, count]
            properties:
              code:
                type: string
                description: 'Стадия конвейера, она же код причины'
              reason:
                type: string
                description: 'Формулировка для читателя, по-русски'
              count: { type: integer, minimum: 1 }
              examples:
                type: array
                maxItems: 5
                items: { type: string }
        windowFrom: { type: string, format: date }
        windowTo: { type: string, format: date }
    TrendReport:
      type: object
      required:
        [id, researchRequestId, version, query, methodologyVersion,
         corpusSnapshotId, coverage, trends, generatedAt]
      properties:
        id: { type: string, format: uuid }
        researchRequestId: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        previousVersionId: { type: [string, 'null'], format: uuid }
        query: { type: string }
        normalizedQuery: { type: string }
        methodologyVersion: { type: string, examples: ['em-1.0.0'] }
        methodologyProfileId: { type: string, format: uuid }
        scoreAggregator: { type: string, examples: ['WEIGHTED_GEOMETRIC'] }
        # Чем посчитан отчёт. Движок методологии выведен из продукта; `methodology` остаётся
        # подписью отчётов, выпущенных до этого, и отсутствие значения у самых старых из них
        # читается так же.
        engine: { type: string, enum: [methodology, signals], default: methodology }
        mode: { type: string, enum: [fast, quality], default: fast }
        corpusSnapshotId: { type: string, format: uuid }
        coverage: { $ref: '#/components/schemas/Coverage' }
        truncated: { type: boolean, default: false }
        # Темы, скрытые из отчёта по пометке смотрящего. Число скрытого лежит в coverage и приходит
        # от движка, имена — отсюда: обратная связь хранит только ключ, а ключ стеммирован
        # («retriev passage»), и показывать аналитику его же решение в таком виде значит требовать
        # расшифровки. Список зависит от смотрящего и поэтому не входит в сам отчёт: он неизменяем
        # и общий, у коллеги с другими пометками скрыто другое.
        hiddenTopics:
          type: array
          items:
            type: object
            required: [trendKey, title]
            properties:
              trendKey: { type: string }
              title: { type: string }
        portrait:
          $ref: '#/components/schemas/DirectionPortrait'
        trends:
          type: array
          items: { $ref: '#/components/schemas/RankedTrend' }
        generatedAt: { type: string, format: date-time }
...
    DirectionPortrait:
      type: object
      description: >-
        Направление в тех немногих величинах, за которые аналитику отвечать вслух. Всё выведено из
        самого отчёта; сгенерированных утверждений здесь нет намеренно — фразу «поле бурно растёт»
        нельзя проверить, а сводка существует для того, чтобы её повторил тот, с кого спросят.


        Оговорки (`lowEvidenceCount`, `unavailableSources`) — обязательные поля, а не приложение:
        оговорка, которую можно не прислать, не будет прислана ровно тогда, когда она важна.
      required: [trendsInReport, candidatesEvaluated, lowEvidenceCount, byLifecycleStage, windowFrom, windowTo, sourcesUsed, unavailableSources, partial]
      properties:
        trendsInReport: { type: integer }
        candidatesEvaluated:
          type: integer
          description: Сколько кандидатов дошло до оценки. «15 из 4000» и «15 из 18» — разные утверждения.
        lowEvidenceCount:
          type: integer
          description: Сколько тем в ТОП держится на тонкой доказательной базе.
        medianFirstMentionYear:
          type: [integer, 'null']
          description: >-
            Медиана, а не среднее: одна тема 1998 года сдвинула бы среднее на годы и не описала бы
            ни одной реальной темы. null для пустого отчёта.
        byLifecycleStage:
          type: array
          description: >-
            Все стадии методологии в её собственном порядке, включая нулевые: отсутствие строки
            «зрелые: 0» читается как «не считали», а порядок по величине менялся бы между отчётами.
          items:
            type: object
            required: [stage, count]
            properties:
              stage: { type: string }
              count: { type: integer }
        windowFrom: { type: string, format: date }
        windowTo: { type: string, format: date }
        sourcesUsed:
          type: array
          items: { type: string }
        unavailableSources:
          type: array
          description: Чего в выводе заведомо нет.
          items: { type: string }
        partial: { type: boolean }
        directionRecognized: { type: boolean, default: true }
    ReportDelta:
      type: object
      required: [entered, left, stayed]
      properties:
        previousReportId:
          type: string
          format: uuid
          description: Отчёт, с которым сравнивали. Отсутствует, если сравнение не выполнялось.
        previousVersion: { type: integer }
        entered:
          type: array
          description: Темы, которых не было в предыдущей версии. Порядок — по рангу в текущей.
          items: { $ref: '#/components/schemas/ReportDeltaEntry' }
        left:
          type: array
          description: Темы, выбывшие из ТОП. Порядок — по рангу в предыдущей версии.
          items: { $ref: '#/components/schemas/ReportDeltaEntry' }
        stayed:
          type: array
          description: >-
            Темы, присутствующие в обеих версиях. Порядок — по росту в ранге, то есть ускорившиеся
            первыми.
          items: { $ref: '#/components/schemas/ReportDeltaEntry' }
        unavailableReason:
          type: string
          description: >-
            Заполнено ровно тогда, когда сравнение не выполнялось, и является частью контракта, а не
            пустым ответом: пустые группы читаются как «ничего не изменилось» — содержательное
            утверждение, противоположное «сравнить не с чем» и «сравнивать нельзя».


            `no-previous-version` — это первая версия отчёта.
            `methodology-changed` — версия методологии изменилась между запусками, и разность баллов
            была бы изменением линейки, а не мира.


            Тип — строка, а не закрытый enum, намеренно: причина показывается пользователю, и
            появление новой не должно ломать уже собранных клиентов. Так же сделаны `stage` и
            `outcome` в `TermExplanation`.
    ReportDeltaEntry:
      type: object
      required: [trendKey, title, rank, score]
      properties:
        trendKey: { type: string }
        title: { type: string }
        rank:
          type: integer
          description: Позиция в той версии, где тема присутствует.
        rankChange:
          type: [integer, 'null']
          description: >-
            На сколько позиций тема поднялась (`ранг_прежний − ранг_текущий`), поэтому
            положительное значение читается как «выросла», хотя ранг считается вниз. Отсутствует
            для тем, представленных лишь в одной версии: прежнего ранга нет, а ноль читался бы как
            «удержала позицию».
        score: { type: number }
        scoreChange:
          type: [number, 'null']
          description: Изменение балла со знаком; отсутствует при том же условии.
    TermExplanationPending:
      type: object
      description: Трассировка ещё считается; ответ будет по тому же адресу.
      required: [state, retryAfterSeconds]
      additionalProperties: false
      properties:
        state:
          type: string
          enum: [RUNNING]
        retryAfterSeconds:
          type: integer
          minimum: 1
          description: Через сколько секунд спросить снова; то же, что в заголовке `Retry-After`.
    TermExplanation:
      type: object
      required: [stages, traces]
      properties:
        stages:
          type: array
          description: >-
            Все стадии конвейера по порядку, как их называет движок. Клиент рисует путь термина по
            этому списку, а не по своей копии: копия молча отстала бы в тот день, когда в конвейере
            появится новая стадия, и продолжила бы выглядеть правильной.
          items: { type: string }
        traces:
          type: array
          items:
            type: object
            required: [term, stage, outcome, inReport]
            properties:
              term:
                type: string
                description: Термин в том виде, в каком его ключевал движок.
              canonicalTerm:
                # 3.1 выражает «может быть null» типом-объединением; `nullable` — синтаксис 3.0,
                # и валидатор 3.1 его не признаёт. Смысл тот же, запись другая.
                type: [string, 'null']
                description: Термин, в который этот был слит при объединении синонимов.
              stage:
                type: string
                description: Последняя достигнутая стадия.
              outcome: { type: string }
              reason:
                type: [string, 'null']
                description: Причина отказа словами, с которыми аналитик может что-то сделать.
              detail:
                type: object
                additionalProperties: true
                description: Числа, решившие исход — пороги и измеренные значения.
              inReport:
                type: boolean
                description: Дошёл ли термин до отчёта.
...